pyproject.toml
[tool.poetry]
name = "school"
version = "0.1.0"
description = ""
authors = ["denisbelenko <denisdel123@>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.4"
psycopg2 = "^2.9.9"
pillow = "^10.3.0"
ipython = "^8.23.0"
djangorestframework = "^3.15.1"
django-filter = "^24.2"
djangorestframework-simplejwt = "^5.3.1"


[tool.poetry.group.dev.dependencies]
python-dotenv = "^1.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
